+++ src/main/java/net/oneandone/sushi/fs/zip/ZipRoot.java
@@ -74,6 +74,11 @@ public class ZipRoot implements Root<ZipNode> {
         if (encodedQuery != null) {
             throw new IllegalArgumentException(encodedQuery);
         }
+        // TODO: springboot's class loader returns resource urls with multiple "!"s like
+        //   jar:file:/Users/mhm/Projects/github.com/net/oneandone/stool/stool/server/target/server-5.0.0-SNAPSHOT-springboot.jar!/BOOT-INF/classes!/stool.version
+        // I didn't find this documented and I didn't find a spec that says this is valid.
+        // So I don't know if that spring specific or in compliance with some specs ...
+        path.replace("!", "");
         return new ZipNode(this, path);
     }