+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ### 3.2.1 (pending) 
 
-* update lazy-foss-parent 1.0.2 to 1.0.3
-
+* update lazy-foss-parent 1.0.2 to 1.2.0
 
 ### 3.2.0 (2018-11-12)