+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * HttpNode tweaks
   * added withHeaders() to configure per-node headers; note that since nodes are immutable,
     the method returns a new modifed node
+  * added withEncodedQuery
+  * added withParameter
 * update lazy-foss-parent 1.0.2 to 1.2.0
 * added Buffer methods:
   * fill(InputStream in, int max);